refactor in progress - switching to laptop
