+++ Assets/Library/src/ui.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6c310ed6c7589d49a2c668289c45790
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
