--- Assets/Library/src/units/Rat.cs
@@ -1,15 +1,0 @@
-﻿namespace Library.src.units
-{
-    [System.Serializable]
-    public class Rat
-    {
-        public string ratName;
-        public float health;
-        public float attack;
-        public float defense;
-        public float speed;
-        public float slowedSpeed;
-
-        //TODO add further fields for pickups, buff, and loading in char model and visuals for each. 
-    }
-}
