--- Assets/Library/src/units/Rat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7f66b83b04539e458b07b1a9d45f68e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
