+++ Assets/Library/src/units/ai.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4af580581b7055a409c008a894226a7a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
