+++ Assets/Resources/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9abbbb34857aaaf41b3b3535cc1ead62
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
