+++ Assets/Visual/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fd7439dc26a5c14e972030e07d38518
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
